tests: move vm update test to a seprate file
This allows scheduling them in a separate job on openQA
